split into separate files and add confirmation dialog with message initializer
